dropped stack implementation
