+++ libraries/tlv-account-resolution/src/lib.rs
@@ -10,7 +10,6 @@
 pub mod account;
 pub mod error;
 pub mod seeds;
-pub mod stack;
 pub mod state;
 
 // Export current sdk types for downstream users building with a different sdk
